get attrssss
